Greatly simplify session creation code.
Uses some helpers to get the data more directly from the python attributes to
the struct. This also removes a lot of the error checking for non python errors.
Basically we will still error out if the attribute is missing or if the value
can't be converted, for anything else, like say None the value just won't be
set in the struct.