Pass proper booleans to callbacks.
This code simply uses 'value ? Py_True : Py_False' as the "O" used with
PyObject_CallFunction will increase the refcount for us, before adding the
arguments to an internal *args (that steals the ref) and the call the
callable with the *args and decref args. Reseting the count that what it
was before our call.