Extract code that converts from PyObject to sp_*_type
This adds helper functions 'bool *_convert(PyObject*, void*)' for
sp_artistbrowse_type, sp_search_type, sp_toplisttype and toplistregion.
These can be used with PyArg_ParseTupleAndKeywords and the format "O&"
to convert the Python object to an approriate C structure.

For all but the toplistregion we convert directly to the right sp_* type.
Regions need a custom struct that contains a sp_toplistregion and username as
we produce both sp_toplistregion and username from the same PyObject.