Use encoded string convter for usernames and passwords
This fixes #103 which was caused by incorrectly encoding an already encoded
string. Ideally this would be fixed with much simpler code in the python API,
but since we expose the C-API login function like we do there is no guarantee
about what we are called with.

The conversion function takes (PyObject*, char**) and allocates space for and
copies the object encoded as utf-8. As such the caller needs to cleanup with
PyMem_Free after this has been used.