Update *_FromSpotify to take an extra bool add_ref.
This extra parameter has been added to everything except session and playlist
folders. Idea is that callers _must_ always care and know if we should be
adding a ref or not. With this change we should be rid of a bunch of lost
libspotify objects that would end up with n+1 refs instead of n.