docs: Support building docs without libspotify
Constants from libspotify that are read by pyspotify already at import time is
cached in a CSV file. Sphinx uses the CSV file to create a mock of
cffi/libspotify with all the constants available. A fab task is used to update
the constant cache whenever libspotify is updated.

This makes it possible to build the API docs on systems that don't have
libspotify, like the Read the Docs builders.